Make PT_XLA_DEBUG analysis's number of frame printed configurable (#5958)
* Make PT_XLA_DEBUG max frame configurable

* lint

* add unit test
